refactor: move ByteUtils related test to ByteUtilsTest
